requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.9.7
django-angular==0.8.2
django-bootstrap3==7.0.1
django-extensions==1.6.7
djangorestframework==3.3.3
inflect==0.2.5
ipython==4.2.0
ipython-genutils==0.1.0
pickleshare==0.7.2
silly==1.1.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
wheel==0.24.0
